refactoring migration of functions from GenericRepository.cs to ShopRepository.cs
