+++ Koop/Models/Repositories/IRepository.cs
@@ -11,9 +11,5 @@ namespace Koop.Models.Repositories
         T GetDetail(Func<T, bool> predicate);
         void Add(T entity);
         void Delete(T entity);
-
-        IEnumerable<Basket> GetBaskets();
-        IEnumerable<UserOrdersHistoryView> GetUserOrders(string firstName, string lastName);
-        Supplier GetSupplier(string abbr);
     }
 }
