+++ Koop/Models/Repositories/IShopRepository.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
+using Koop.models;
 using Koop.Models.RepositoryModels;
 
 namespace Koop.Models.Repositories
@@ -15,5 +16,10 @@ namespace Koop.Models.Repositories
     {
         IEnumerable<ProductsShop> GetProductsShop(Expression<Func<ProductsShop, object>> orderBy, int start, int count,
             OrderDirection orderDirection = OrderDirection.Asc);
+        
+        IEnumerable<Basket> GetBaskets();
+        IEnumerable<UserOrdersHistoryView> GetUserOrders(string firstName, string lastName);
+        Supplier GetSupplier(string abbr);
+        IEnumerable<Order> GetBigOrders();
     }
 }
